+++ src/main/java/factorisation/AppView.java
@@ -1,3 +1,5 @@
+package factorisation;
+
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.geometry.Insets;
