+++ src/main/java/factorisation/Factoriser.java
@@ -1,3 +1,5 @@
+package factorisation;
+
 import java.util.Vector;
 
 public class Factoriser {
