+++ src/main/java/factorisation/Main.java
@@ -1,4 +1,4 @@
-/*
+package factorisation;/*
  * Ernestas Seminogovas
  * Prime factorisation with JavaFX GUI
  */
@@ -9,12 +9,7 @@ import javafx.stage.Stage;
 
 public class Main extends Application {
 
-    public static void main(String args[]) {
-        launch(args);
-    }
-
     @Override
-
     public void start(Stage primaryStage) throws Exception {
         AppView view = new AppView();
         view.setupUI(primaryStage);
