+++ src/test/java/InputParserTest.java
@@ -1,13 +1,9 @@
-import javafx.application.Application;
+import factorisation.InputParser;
 import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.mockito.Mock;
-
-import javax.xml.soap.Text;
 
 import static org.junit.Assert.*;
 
@@ -33,9 +29,9 @@ public class InputParserTest{
         fieldC.setText("null");
         fieldD.setText("10");
     }
-    @Test
+    @Test(expected = NumberFormatException.class)
     public void parseInputA() {
-        assertEquals(-1,InputParser.parseInput(fieldA));
+        assertEquals(-1, InputParser.parseInput(fieldA));
     }
 
     @Test(expected = NumberFormatException.class)
