+++ src/test/java/MainTest.java
@@ -1,7 +1,6 @@
+import factorisation.Main;
 import javafx.stage.Stage;
 import org.junit.Test;
-
-import static org.junit.Assert.*;
 
 public class MainTest {
 
@@ -9,9 +8,4 @@ public class MainTest {
     public void main() {
     }
 
-    @Test
-    public void start() throws Exception {
-        Main main = new Main();
-        main.start(new Stage());
-    }
 }
